Finish tidying up the frontend/lib directory (#1169)
This finished the work started in #1168 by creating a few new thematic directories for front-end modules:

* `progress` contains files related to the concept of progressing through a wizard-like flow.
* `networking` contains files related to loading resources over the network, and presenting the user with loading screens and such.
* `ui` contains files related to the user interface that aren't covered by any of the other thematic directories, e.g. reusable widgets and such.
* `util` contains various utilities that aren't covered by any of the other thematic directories and can include constants, miscellaneous React Hooks and other reusable code.

At this point the only actual files in `frontend/lib` are ones that are the entry points of the app and/or control global state in some way.

Note that the folders in `frontend/lib` are not intended to be mutually exclusive categories; there will definitely be times when we'll question whether a module belongs in one folder versus another, and I think that's OK.  The good thing is that for now this hopefully makes the directory less overwhelming and easy to get lost in, and we can always reorganize things in a different way later.

I also opted to have a shallow directory hierarchy rather than a deep one. For example, I could have put `forms` inside the `ui` folder, since a form is a type of user interface, but I decided not to because I've found that deep nesting can get confusing.  We can always revisit this decision in the future, though.