Add prettier integration for TypeScript files (#1170)
This adds prettier integration for TypeScript files, and tests during CI to make sure everything is pretty.  We can add more file types later!

A section called "Prettier" has been added to the README which documents how to use it, which is essentially to run `yarn prettier:fix` to fix everything before pushing to GitHub.