Factor out files from styles.scss. (#1178)
This factors out some files from `styles.scss` which should have been factored out a while ago.  I'm doing it now so it will be easier to reuse these styles from the NoRent site if we want.

Note that I'm not 100% certain that styling will be preserved, because I'm effectively re-ordering the CSS, so there may be some subtle styling changes.  I manually test the site to ensure that everything looks okay, though.