Use node-sass-chokidar. Fixes #300. (#1179)
This fixes #300 by using node-sass-chokidar.

I'm not a big fan of what this is doing to our `yarn.lock`, as it seems to have brought in yet another version of chokidar and a few other dependencies we already have, but since it's all on the build side and not in the browser I think it's ok.  Still, we should figure out if there's an easy way to clean up the `yarn.lock` at some point, so I've filed #1180 to address it later.