Add /dev/ to NoRent.org, add AppContext.siteRoutes. (#1183)
This makes the `/dev/` routes accessible on NoRent.org, which makes it easy to see the site's URLs and visit its example pages.

Doing this required decoupling the development routes from the JustFix site's `Routes` object.  To make this less annoying, I added a `siteRoutes` property to the `AppContext` type which represents all routes common to both the JustFix and NoRent sites, which the development routes are included in.  This allows the development routes to easily "attach" themselves to either site.  In the future, I'm hoping it will make it easy for any route, regardless of site, to easily do common things, like redirecting the user to a login page if they're not logged in.