Make the <Navbar> component generic, add it to NoRent.org. (#1185)
This makes the `<Navbar>` component more generic; it still contains links to a few common areas like the admin and the dev tools, but otherwise it leaves out the branding and any extra links.

A new `<JustfixNavbar>` adds in all the JustFix site-specific branding and links.  Right now the NoRent.org site uses the default `<Navbar>` but we can add its branding and extra links later.

Currently the CSS for the navbar is unchanged, which means that it looks rather out-of-place on the NoRent.org site. We'll likely need to refactor the SASS to not necessarily make it the color of the primary color, since NoRent.org won't use that approach... Oof.