Add NoRent.org model scaffolding (#1187)
We don't know exactly how we're going to store data about non-NYC users in the database, so until we do, we'll scaffold our model by just storing everything in the session.  This allows us to build the front-end and even some of the back-end functionality without having to block on figuring out the user account and database schema details.

One thing to keep in mind is that because the NoRent scaffolding is part of the `session` GraphQL query, and because it's explicitly queried as part of `AllSessionInfo.graphql`, we need to be careful about evolving it in a way that breaks backwards compatibility: since even regular JustFix.nyc users will be querying this information (even though they don't actually use it), a backwards-incompatible change such as removing a field could result in out-of-date clients inadvertently crashing. (Obviously a simple page reload would fix this, but users may not know to do that.)