Add NoRent letter pdf/html placeholder page (#1200)
This builds on #1194 by allowing static pages to render as PDF pages.  It also adds a new `<LetterStaticPage>` component that includes the same CSS as the letter of complaint, and makes it easy to quickly add routes for different kinds of letters.

Finally, it uses this new tooling to add a NoRent letter PDF/HTML placeholder page.

The CSS styles are actually retrieved via GraphQL for now because it was the most expedient option given our legacy code for rendering PDFs via Django templates.