Factor out lambda-response-http-headers.tsx. (#1211)
For the past few days I've noticed that whenever I modify a file, it often causes Jest to re-run _all_ tests.

It seems the ultimate cause of this is #1200, which added a new `LambdaResponseHttpHeaders` type to `lambda.tsx`, which a few modules imported.  As far as I can tell, it seems like Jest treats type imports the same way it treats implementation imports (which is funky because it doesn't actually do any type checking), so this ultimately caused a change to almost any file to result in re-running the whole project's test suite.

This moves the type into a separate file, thus breaking the cycle of misery.

Figuring out the culprit was done via the following command:

```
node_modules/.bin/jest --findRelatedTests --listTests frontend/lib/app.tsx
```

This listed all the tests that would be run if `app.tsx` was changed.  I then did a manual `git bisect` (I really need to learn how to actually use that command someday) to find when the list of tests exploded, which allowed me to narrow down the offending commit to #1200.