Add create account page (#1233)
This adds a create account page to the NoRent letter builder.  Existing users will now be fast-forwarded through to the landlord management info page (this actually ignores a number of edge cases, for which #1234 has been filed).  We also use the user's `JustfixUser` and `OnboardingInfo` model data to populate relevant parts of letter content, rather than the `norentScaffolding`.