Add support for intra-page links. (#1239)
This adds support for intra-page links, e.g. `<a href="#some-section">`, as well as support for links to anchors on other pages.  It smoothly scrolls when users click on links.

Currently it doesn't support use cases where the page of the target link needs to be lazy-loaded, however.  But this won't be a concern for the NoRent site, where we need it now.

Also, in order to make a link scroll to its target even if the URL is currently already targeted there, we'll need to use a newly-added `<ScrollyLink>` class for now.  In the future we might be able to make any standard `<Link>` work but for now I didn't want to risk global code or a third-party scrolling library interfering with our existing scroll logic.