Add RentPeriod and Letter models to norent. (#1240)
This adds the beginnings of backend logic for sending of the NoRent letter and querying the sending status.

Unlike the letter of complaint, the NoRent letter isn't a one-to-one model with the user: instead, we've got the concept of a "rent period", which has a concept of a "payment date".  Only one letter per user can exist per period, so e.g. once a user has sent a letter for the one rent period, they can't send another until the next period begins.  (And by "begins" we currently mean "exists in the database".)

This also adds GraphQL endpoints for querying the current rent period and retrieving information about the user's most recent letter.  It also adds a mutation to send a letter, but it's only partly complete, as finishing it will take more work.