Add render_raw_lambda_static_content(). (#1241)
This adds a new function at `project.views.render_raw_lambda_static_content()` that allows the server to obtain the result of a static page rendered by the React front-end (via functionality introduced in #1194).  We'll need this to complete the work begun in #1240, so we can generate the PDF HTML from our GraphQL mutation for sending the norent letter.