Ask for LL name and contact types (#1253)
This uses the GraphQL mutation added in #1252 to ask the user for their landlord name and what kinds of contact info they have for them.  It also replaces the massive landlord details scaffolding form with the three separate steps from the mockups, and includes support for NYC users who had their landlord details looked up in open data.

This does introduce a few wonky edge cases, though: once a user has entered an email and/or mailing address for their landlrod, they can change it, but they can't _remove_ it.  For instance, if they say they only have an email address, go forward and enter one, then go back and say they only have a mailing address, enter one and then send the letter, the letter will be sent to _both_ the email address they previously entered _and_ the mailing address.  This might not be their intent.  I'm going to file this as an issue to address later.

There's also a few DRY violations in this code, which I'll address later.