Fix login bugs involving hard redirects and forced refreshes. (#1273)
Oy, after having lots of weird issues logging into both my local dev setup and the demo site, I realized we have a few bugs involving login:

1. When any user logs in, we have a function called `performHardOrSoftRedirect()` that checks to see if the given URL is one that can be redirected to as part of the SPA.  However, it was being passed an absolute URL, and until now, it expected any local passed-in URLs to be passed in without any protocol or host information.  This resulted in a hard redirect to _always_ occur.

2. When admin users logged in, we should always force a page reload, to disable third-party analytics.  However, this often ended up being jarring and sometimes occurred just before a page transition, which resulted in either the page transition not happening, or the transition _appearing_ to happen but then immediately going back to the old page (at least, I think this is what was going on).

This fixes both bugs.  To solve (2), we wait a second before presenting the admin user with a modal that lets them know a page reload is about to occur, with an explanation for why.