Add session.norentLettersSent GraphQL field. (#1287)
This adds a `norentLettersSent` GraphQL field that we can use to populate a "letters sent" area on the NoRent site.  While this will likely only be used on the splash page, it's easiest to just add it to the session object for now, since (A) that's easiest for the front-end and (B) we don't want to require the splash page to double-render on the server side by requesting another GraphQL query.  Besides, we can always improve the performance of the field through smart caching in the future, which should be easy since this field is intended for "social proof" and its precise accuracy doesn't really matter (as long as we don't over-count, which is impossible).