Add support for legacy tenants app users in NoRent. (#1290)
Legacy tenants app users (that is, users of the ancient "build your case"-based app at  JustFixNYC/tenants) have special "barebones" accounts on the new app that is used exclusively to forward them to the old system as soon as they login.  This functionality is currently documented in the `legacy_tenants` directory of this repository.

While these users have a related `LegacyUserInfo` model with a `prefers_legacy_app` boolean that tells us whether to forward the user to the legacy tenants app, there's not really any way for the user to use any of our services on this new platform unless we manually create `OnboardingInfo` for them.  And it's hard to reuse our existing onboarding flow for them because it assumes the user doesn't have a `JustfixUser` model yet, which these users do.

Furthermore, the password of these users is stored in, and validated by, the legacy system, which can't be maintained indefinitely as that system will reach end-of-life soon.

So the solution this PR takes is to simply provide an automated, opt-in way for such users to "migrate" their account to the new system.  Once the user enters their phone number on NoRent.org, they will see a screen that informs them about the migration process.

Once they click "migrate my account", their existing "barebones" account will be assigned a new random phone number and made inactive (a better alternative than permanent deletion, in case we want to undo anything) and they will be able to create their new account.  The phone number they've already entered into the system will be used for the new account, so they'll essentially be treated from this point forward as though they're a brand new user (the next question they'll be asked, for instance, is what their first and last name is).