FIXED the ridiculous issue with our unclickable sticky menu on safari mobile (#1311)
After a lot of pain and turmoil, this PR finally fixes an issue that made the iOS safari navbar consistently block our sticky footer button so that users had essentially no way of clicking on it. Unfortunately, this fix does introduce some added Jank, as it seems Safari doesn't play well with "sticky" positioning in a smooth way. Namely, if a user scrolls up or down vigorously, the sticky button menu will have a split-second of lag time in which It tries to keep up with the scrolling movement.

I ran this issue by Tahnee and she is ok with it for launch, but suggests that we circle back to it when we go through other styling updates post-launch. Will file an issue to keep track of this.