Make 'cancel' button on NoRent builder clear session. (#1318)
This makes the "cancel" button at the beginning of the NoRent letter builder flow clear the session, like the ones on the JustFix.nyc onboarding flows do.

The existing `<ClearSessionButton>` assumes it's being put on a page next to a different form's buttons, and therefore involves all kinds of React Portal acrobatics. However, we don't need such sophistication on the NoRent letter builder's welcome page because it doesn't actually have a form to fill, so I made a separate `<SimpleClearSessionButton>` that doesn't require all that portal goop.