Add special flow if user is in state w/o protections. (#1316)
If the user is in a state without protections, we need to inform them during the KYR step and let them know that they can create an account and we'll inform them when their state has better protections, but they can't send a letter right now.  We also link out to the FAQ section on "STATES WITH LIMITED PROTECTIONS" so the user can learn more.