Show error pages if user isn't in proper session state to view them (#1322)
This decorates a bunch of our letter builder steps with various checks that ensure the user has a session state in which it actually makes sense for the user to see them.  If it doesn't make sense, we show alternate content with a call to action that points them in the right direction.

Broadly speaking we have a few kinds of steps:

* **Onboarding steps**, where the user is entering information before they've actually created an account.  In these steps it's actually _bad_ for a user to be logged-in, so we want to direct them to the latest step they've completed.

* **Not sent letter steps**, where the user has created an account but not yet sent a rent letter for the current rent period.  For these steps, if the user isn't logged-in, we want to direct them to log in; and if they are logged in but have already sent a letter, we want to direct them to the confirmation page.

Aside from that there are a few other edge cases which this PR attempts to handle.