Replace optInToRttcComms with norentOptInToRttcComms. (#1330)
Oof, in #1329 I forgot that opting-in to RTTC communications needs to be done both by brand-new users to NoRent.org, as well as users arriving from JustFix.nyc with existing accounts.   Since the best place to store  temporary onboarding details about NoRent users is in `norentScaffolding`, I moved the mutation over to that package and renamed it to `norentOptInToRttcComms`.