Add progressively-enhanced <CityAndStateField>. (#1360)
This wraps the mapbox autocomplete widget introduced in #1205 within a progressively-enhanced component that shows the regular city/state fields if mapbox support is disabled, or if mapbox is unavailable for some reason.  It also substitutes the city/state fields in the NoRent letter builder flow with this new component.

Finally, it also adds a new `FormFieldPropsBuilder` class that makes testing the component, and other components like it, easier.