Add tests for deploy.py. (#1363)
I'd like to start making some improvements to our deployment script to support things like Heroku preboot, but that introduces enough complexity that I'd like to add proper unit tests for our deploy script first.

This also adds a new `Snapshot` helper class that makes it easier to do Jest-like snapshot testing using pytest.

This also removes `deploy.py` from our `.coveragerc` so we're now tracking coverage of it.  The coverage for the whole project is going down right now because we don't have enough tests to cover the whole thing, and I need to move on to I18n for now so I don't have time to add full test coverage, so we'll just have to do with what we've got for now.