Jump straight to top of page for some page transitions. (#1369)
For a whitelist of pages, this jumps the user straight to the top when a push/replace page transition occurs, which is preferable for pretty much any page that isn't transitioned-to via a slide animation.  It also seems to prevent a bug on Firefox and some other browsers where, if a slide transition doesn't occur, the smooth-scroll-to-top doesn't occur either, and as a result the user may be left stranded in the middle of the page that just got navigated to.

This fix definitely isn't ideal; we should probably instead implement something that _only_ smoothly scrolls to the top when a gentle transition between the pages is occurring, but right now this is the most expedient and safest option.  I've filed #1372 to address it later.