Refactor partner links on NoRent KYR page to be more DRY (#1377)
This PR implements a new component <OneOrTwoPartnerLinks> for our "Know your rights" page on the NoRent Letter Builder flow that dynamically generates a list of two links to partners (if a local state partner is provided) or just one link to the default national partner. This helps us render copy for this page in a more DRY fashion, and fixes #1335.