Add Lingui support. (#1376)
This adds front-end localization via the Lingui framework.  For now I've just localized two strings in the NoRent.org site to Spanish, to make sure that everything works.  We're using code splitting to only load the message catalog for the currently selected locale, but outside of that, unfortunately, we have one catalog for the entire codebase, which is suboptimal.  I have some ideas for working around this in the future but we'll deal with that if/when the one massive catalog becomes a burden.

Because we don't want the Django server redirecting browsers that prefer Spanish to our 99.9% incomplete localization, I've added an `ENABLE_WIP_LOCALES` setting that needs to be enabled in order for developers to access the Spanish localization.