Add support for i18n in norent airtable. (#1392)
This adds some plumbing for i18n to our code that pulls from the norent airtable.  While it doesn't fully enable i18n in the app, it now knows to treat any field name that ends in " (English)" as being localized to `en` and anything ending in " (Spanish)" as being localized in `es`.

At present the code in this PR is only paying attention to the English version, and it's generating the exact same JSON as it did before.  This allows translators to start localizing the Airtable ASAP.

In a future commit, we'll create separate JSON files for each table in each language.  Then we'll load the correct one at runtime.