Load locale-specific national metadata bundle. (#1400)
This loads the locale-specific national metadata bundle for the NoRent site, which exposes the NoRent localized airtable data on the front-end, completing the work started in #1392.  As with the Lingui catalogs, we use code splitting to ensure that only the bundle for the current locale is retrieved.

One unfortunate aspect of this approach to localizing our national metadata is that we've got a lot of data in the locale-specific bundles that isn't actually locale-specific, and that comes with a few wrinkles, which I've documented in #1401.