Factor out locale-choices.json (#1402)
Until now, our front-end and back-end kept track of available locales in different ways, but this PR unifies the data in a new `locale-choices.json` file.  The back-end accesses this from a new `project.locales` module and the front-end accesses it from a new `locale-choices.ts` module that is auto-generated by commondatabuilder.