Refactored FAQ content to make all answer preview fields optional (#1398)
This PR simplifies our NoRent FAQ content through refactoring the array we store it in. The main difference here is that we now update the Faq type to allow for some fields to be optional.
The consequences of this change are that:
Only FAQ entries that we want to include in a preview section (like on the homepage or "The Letter" page) will need to have a numeric priority and answerPreview content defined.
The priorityInPreview field will determine the order they show up on the preview section.
For the full list of FAQ entries on the FAQs page, order will be determined by the order entries appear in the FAQ content array, once they are sorted into categories. Therefore, adding in a new FAQ entry to this page simply means adding a new element within the array— no hassle with defining a "priority" and shifting other entries' priorities to line up.
My hope is that these changes will set us up to have a smoother time with internationalization