Split compiled message catalogs into multiple files (#1407)
This is an attempt to resolve one of the big pain points present in lingui, which is that lingui only creates a single monolithic message file.

Right now this is solved by a script called `localebuilder.js` that parses the `messages.po` created by `lingui extract` and the `messages.js` created by `lingui compile` and attempts to associate AST nodes representing message definitions with the source files they're present in, as documented in the references of `messages.po`.

The JS is parsed by Babel, which is already a dependency, while the PO is parsed by pofile, a library that lingui already uses, so in the end we don't bring in any new dependencies.

We then split up the `messages.js` by slicing it up into "chunks": right now just a "base" chunk and a "norent" one, putting all the messages that come from `frontend/lib/norent` into one compiled message catalog and everything else into another.  This leaves it up to us to make sure that anything in our code which uses components from that directory also loads its corresponding message catalog, which isn't awesome, but it's still way better than having a single monolithic bundle.