Factor out browser-redirect.tsx. (#1410)
This factors out `browser-redirect.tsx` to contain all code related to redirecting the browser.  It also makes its `performHardOrSoftRedirect` function work if called on norent.org.

It also renames `routes.ts` to `justfix-routes.ts`, and its default export from `Routes` to `JustfixRoutes`.  This should make it easier to discern what code is justfix.nyc-specific, versus generic to justfix.nyc and norent.org.

Finally, it adds a _new_ `routes.ts` that consists of the `AppSiteRoutes` type from `app-context.tsx`, along with a new helper function, `getGlobalSiteRoutes()`, which allows functions that don't have access to an `AppContext` to retrieve the site routes.