Improve force-refresh on admin login. (#1411)
This improves the janky "let's show a dialog telling the admin that the page will refresh" behavior from #1273 by just ensuring that the next page redirect (which is imminent) will be a hard one.  It also tees us up for reloading the page if we know a new version of front-end code is available.

I originally thought that I could simply change the `forceRefresh` prop passed to the `BrowserRouter` to do this, but it looks like that prop is only useful when the component is first created, as it ultimately is used to instantiate the router's `History` instance, and that object keeps `forceRefresh` private and read-only after construction.