Make it easier to test GraphQL with AppTesterPal. (#1419)
This adds a two new methods to our `AppTesterPal` class, `withQuery()` and `withFormMutation()`, which make it much easier to test our GraphQL code in a type-safe way.  This renders some of the pre-existing methods for testing GraphQL code superfluous, so those have been removed.

It also replaces the `getFirstRequest()` method with `getLatestRequest()`, which is more useful for testing.