Ensure signup intent test doesn't randomly fail. (#1427)
I noticed that the signup intent test was randomly failing, and it looks like it's because it was wrapping a dynamic code import in `friendlyLoad`, which ensured that the import took at least one second (intended to result in non-janky loading screens).  Aside from being an unnecessary delay in tests, it looks like it also resulted in the test's `waitFor` to time out intermittently.  This fixes the problem and speeds up the test by altering `friendlyLoad()` to never delay imports when running tests.