Add Amplitude integration to front-end. (#1439)
This attempts to add amplitude integration to the front-end, so we have a bit more visibility into what our users are doing.

As per Amplitude's Data Taxonomy Playbook, I've attempted to use event names that are concise without being too broad:

* **"Submitted form without errors"** is a form submission on any site without errors.
* **"Submitted form with errors"** is a form submission on any site with errors.
* **"Visited [something] page"** is a page view on a particular page such as JustFix Letter of Complaint, NoRent letter builder, and so on.

Right now this is entirely uncommented and untested because I just want to see if it's even useful before doing more.

Various user properties are also reported, such as what state/city the user claims to be from.  No sensitive user data or PII is submitted to Amplitude.