Ensure localization catalogs don't go out of date. (#1443)
This fixes #1442 by checking to make sure the content of localization catalogs never goes out-of-date.  Note that it does _not_ make sure that file/line number references are accurate, since that's likely to change frequently and doesn't affect the accuracy of our translations (although it might be confusing for translators, I don't think we should consider such inaccuracies a broken build).

Unfortunately, though, this _does_ have a CI impact; on my system the new `yarn lingui:extract-and-check` command takes about 18 seconds to run, primarily because it needs to run `yarn lingui:extract`.