Rename JustfixRoutes.onboarding to .locOnboarding (#1457)
This started out as a PR to tell Amplitude to classify `/onboarding` as being part of LOC, as it was created in a time before we had a notion of signup intent.  But then I realized the name `JustfixRoutes.onboarding` was actually misleading, and that it was really LOC-specific onboarding, so I renamed it.

So, this both instructs Amplitude accordingly and renames things to make more sense.