Fix dates in sample letter. (#1460)
Oof, in #1456 I forgot that the "today's date" showing in the NoRent letter uses the current date, which makes the snapshot test I added extremely brittle. This changes the test to use a hard-coded date.

I also realized that the sample letter static page's payment date is May 1, which at this point is out-of-date, which could mislead observant users into thinking that they'd be sending a letter for May 1. This fixes the page to prefer the latest payment date instead.