Determine whether user is in LA in backend. (#1462)
We do the LA determination on the front-end right now.  This has some disadvantages:

* LA has about 400 zip codes, so it kind of needlessly bloats our JS bundle.

* The determination we're trying to make can't actually be determined solely by zip codes, so even our current approach isn't perfect.  Ideally we'd do some GIS work, taking into account the user's lat/long, to accurately figure out if they're in LA.  Doing this in the front-end would needlessly bloat our JS bundle even more.

* Because it's JS code that runs on the front-end, it's hard to build a database query that filters all LA users in the back-end, which is something we've been asked for.

So this resolves the issues by adding a `norentScaffolding.isInLosAngeles` GraphQL field that makes the determination on the back-end, and removing the functionality from the front-end.  It currently works the same way the front-end functionality did, but in the future we can improve it to be more accurate.