Add a --garble option to localebuilder. (#1468)
Well I probably shouldn't have done this, but I was so loathe to manually find un-internationalized text in our codebase that making this was more fun, so here we are.

When one runs `node localebuilder.js --garble`, the non-English message catalogs will be replaced with "garbage" translations, which makes it easy to see what's not localized in the app.  

Obviously the garbage translation should never actually be committed.

The tricky part about this was preserving Lingui's React tags and ICU MessageFormat syntax.  While it's code that will only ever be used during development, the only reasonable way forward was by writing the tests first, and then making them pass (TDD).