Add 'locale' field to JustfixUser model. (#1472)
This helps with #1384 by adding a `locale` field to the `JustfixUser` model.  This will help us when we need to initiate tasks on behalf of the user from a context where we can't retrieve their email from a request, e.g. sending them an email or text message from a worker process.

The `locale` field is set when a user first onboards with us, and its value is taken from the request, so if the user onboards in Spanish, `locale` will be `es`.