Install full ICU data to tenants2_base image. (#1494)
This is an alternative to #1492 that installs full ICU data in the base Docker image, `tenants2_base:0.10`,  resulting in better dev/prod parity.

While it does add 27 megabytes to the base image, it also frees up about 11 by removing the pip cache, which it didn't do before.  Combined with #1493 we still have a pretty big net space savings.

It also adds a test to ensure that the `Intl` API works as expected on the server.