Redirect Spanish routes on JustFix.nyc to the English version. (#1498)
NoRent.org is localized but JustFix.nyc isn't; our server's locale negotiation logic doesn't know this, so it might send the user to the Spanish version of JustFix.nyc (e.g., from `/` to `/es/`), and if that happens we want to redirect the user to the English version, since the Spanish version of JustFix.nyc isn't ready yet.

We're just doing this on the client-side for now because we don't really know what our localization strategy will be for JustFix.nyc: one idea that's been floated is to actually localize our offerings product-by-product, e.g. localizing EHPA before LOC, which would mean we'd need to granualize this redirection based on knowledge of client-side routes, so the client-side seems the best place to put this logic for now.