Add a language toggle to NoRent. (#1503)
This adds a language toggle to the footer of NoRent.  I put it in the footer instead of the navbar because the navbar was becoming quite cluttered on desktop, and on mobile it'd be hidden behind a hamburger menu; it's also based on a similar placement on websites like mozilla.org.

It's not exactly highly visible but my hope is that Django's auto-redirection of browsers to the appropriate locale based on their language preferences will take care of sending most users to the appropriate locale.  (That is, if a user's browser is in Spanish and they go to https://norent.org/, they will be transparently redirected to https://norent.org/es/.)

Furthermore, it seems like this kind of solution isn't necessarily mutually exclusive with putting some kind of UI at the top of the page: as with the navbar links, we can have them both at the header and footer.