Show safe mode UI if app fails to load after 10s. (#1514)
Fixes #809.

I noticed while working on #1513 that there were some weird situations in my development environment in which the main scripts failed to load (or at least failed to initialize the React app) and this made the navbar language toggle menu do nothing, since it requires JS.

This adds a fail-safe mechanism such that if `window.SafeMode.appIsReady()` isn't called within 10 seconds, the safe mode UI is displayed.  The function is called immediately after the app is initialized, so this shouldn't ever happen as long as the user's internet connection is working and our JS bundles aren't massive.

Additionally, if Rollbar is enabled, an error should be logged via it (since the mechanism works by throwing an uncaught exception), so we should be able to see how frequently this gets triggered.