Apply snazzy checkmark to all Tenant Platform confirmation pages. (#1522)
Fixes #1152.  In doing so, it allows the `withHeading` prop of the `<Page>` component to be a render prop that takes the page title and returns JSX for the heading.  It also adds a new `renderSuccessHeading` function that can be used for this render prop, to render the heading with the snazzy checkmark.

Note that this does *not* address #1269, i.e. NoRent and JustFix still use different markup/CSS for styling their checkmark headers.  The work done in this PR will still help towards resolving that, though!