Add NY-specific KYR links. (#1510)
This adds New York-specific outbound links to tenant resources in the NoRent letter builder KYR page.

It also adds a new `<LocalizedOutboundLink>` component that should make it easier for us to add links to localized resources in the future, i.e. where the URL of what we're linking to may be different depending on the locale.