Create a stub React-rendered LOC. (#1533)
This factors out a new `letter-content-util.tsx` module, which contains letter rendering code common to both the NoRent letter and the Letter of Complaint.  We then use this functionality to render a stub sample React-rendered Letter of Complaint at `/loc/sample-letter.pdf`.

Note that the "real" LOC used by the platform is still done on the Python-side: this just starts the process of moving the LOC's rendering to the React-side, which will make internationalizing it much easier (#1508).