Move <title> tag up in <head>. (#1540)
Axel from Whole Whale mentioned that they noticed a significant amount of landing pages being labeled as "not set" and realized that what they had in common was the title tag’s location. Apparently if it is not located before the Google Tag Manager rendered tracking tag, it will usually not be read on the first page view on site. He said that for norent.org and app.justfix.nyc, the title tag just needs to be moved up since it is currently the last element inside the head.

This implements that change, so hopefully GTM will work better!