update moratorium faq url in all instances across site (#1543)
This small PR simply updates all instances of our link out to the RTC Eviction Moratorium FAQs resource. It also saves the URL as a constant to make things more DRY. We could create a whole env variable to make things even easier, but since this is a temporary UI fix, I think that is overkill.