Add mailchimp integration, CORS endpoint. (#1549)
This adds a `mailchimp_subscribe` command that allows email addresses to be subscribed to Mailchimp.  It also adds a CORS endpoint at `/mailchimp/subscribe` that WoW and the org site can POST to in order to subscribe folks.  Performing a GET to this endpoint in the browser will show documentation on the endpoint.